+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/pipeline/VertexProgramBuilder.java
@@ -9,12 +9,13 @@ import com.dynamo.bob.BuilderParams;
 import com.dynamo.bob.CompileExceptionError;
 import com.dynamo.bob.Task;
 import com.dynamo.bob.fs.IResource;
-import com.dynamo.bob.pipeline.ShaderProgramBuilder;
 import com.dynamo.bob.pipeline.ShaderUtil.ES2ToES3Converter;
 import com.dynamo.graphics.proto.Graphics.ShaderDesc;
 
 @BuilderParams(name = "VertexProgram", inExts = ".vp", outExt = ".vpc")
 public class VertexProgramBuilder extends ShaderProgramBuilder {
+
+    private static final ES2ToES3Converter.ShaderType SHADER_TYPE = ES2ToES3Converter.ShaderType.VERTEX_SHADER;
 
     @Override
     public void build(Task<Void> task) throws IOException, CompileExceptionError {
@@ -22,20 +23,9 @@ public class VertexProgramBuilder extends ShaderProgramBuilder {
         try (ByteArrayInputStream is = new ByteArrayInputStream(in.getContent())) {
             boolean isDebug = (project.hasOption("debug") || (project.option("variant", Bob.VARIANT_RELEASE) != Bob.VARIANT_RELEASE));
 
-            ShaderDesc shaderDesc = compile(is, ES2ToES3Converter.ShaderType.VERTEX_SHADER, in, task.getOutputs().get(0).getPath(), project.getPlatformStrings()[0], isDebug);
+            ShaderDesc shaderDesc = compile(is, SHADER_TYPE, in, task.getOutputs().get(0).getPath(), project.getPlatformStrings()[0], isDebug);
             task.output(0).setContent(shaderDesc.toByteArray());
         }
-    }
-
-    public Task<Void> create(IResource input) throws IOException, CompileExceptionError {
-
-
-        Task.TaskBuilder<Void> taskBuilder = Task.<Void>newBuilder(this)
-            .setName(params.name())
-            .addInput(input);
-        taskBuilder.addOutput(input.changeExt(params.outExt()));
-
-        return taskBuilder.build();
     }
 
     public void writeExtraDirectives(PrintWriter writer)
@@ -49,8 +39,9 @@ public class VertexProgramBuilder extends ShaderProgramBuilder {
     }
 
     public static void main(String[] args) throws IOException, CompileExceptionError {
+        System.setProperty("java.awt.headless", "true");
     	VertexProgramBuilder builder = new VertexProgramBuilder();
-        builder.BuildShader(args, ES2ToES3Converter.ShaderType.VERTEX_SHADER);
+        builder.BuildShader(args, SHADER_TYPE);
     }
 
 }
